+++ docs/src/app/(private)/experiments/accordion/horizontal.module.css
@@ -14,7 +14,7 @@
   box-shadow: var(--shadow);
   background-color: rgb(0 0 0 / 0.12);
   border-radius: 0.3rem;
-  height: 40rem;
+  height: 25rem;
   display: inline-flex;
   flex-direction: column;
